Move some things around a bit, delete session workers
